Move Service mocks to Test Utilities
