+++ test/NuGet.Clients.Tests/NuGet.PackageManagement.VisualStudio.Test/CpsPackageReferenceProjectTests.cs
@@ -53,7 +53,8 @@ namespace NuGet.PackageManagement.VisualStudio.Test
                 { constant.FlightFlag, true },
             };
 
-            var service = new NuGetExperimentationService(new TestEnvironmentVariableReader(new Dictionary<string, string>()), NuGetExperimentationServiceUtility.GetMock(flightsEnabled).Object, new Lazy<IOutputConsoleProvider>(() => new TestOutputConsoleProvider()));
+            var service = new NuGetExperimentationService(new TestEnvironmentVariableReader(new Dictionary<string, string>()), new TestVisualStudioExperimentalService(flightsEnabled), new Lazy<IOutputConsoleProvider>(() => new TestOutputConsoleProvider()));
+            //var service = new NuGetExperimentationService(new TestEnvironmentVariableReader(new Dictionary<string, string>()), NuGetExperimentationServiceUtility.GetMock(flightsEnabled).Object, new Lazy<IOutputConsoleProvider>(() => new TestOutputConsoleProvider()));
 
             service.IsExperimentEnabled(ExperimentationConstants.TransitiveDependenciesInPMUI).Should().Be(true);
             componentModel.Setup(x => x.GetService<INuGetExperimentationService>()).Returns(service);
