+++ test/NuGet.Clients.Tests/NuGet.VisualStudio.Common.Test/NuGetExperimentationServiceTests.cs
@@ -3,13 +3,10 @@
 
 using System;
 using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Windows.Media;
 using FluentAssertions;
-using Microsoft.VisualStudio.Experimentation;
 using NuGet.Common;
 using Test.Utility;
+using Test.Utility.VisualStudio;
 using Xunit;
 
 namespace NuGet.VisualStudio.Common.Test
@@ -192,110 +189,6 @@ namespace NuGet.VisualStudio.Common.Test
             OutputMessages.Should().ContainMatch($"*{forcedOffExperiment.FlightFlag}*{forcedOffExperiment.FlightEnvironmentVariable}*0*");
             OutputMessages.Should().ContainMatch($"*{forcedOnExperiment.FlightFlag}*{forcedOnExperiment.FlightEnvironmentVariable}*1*");
             OutputMessages.Should().NotContainMatch($"*{noOverrideExperiment.FlightFlag}*{noOverrideExperiment.FlightEnvironmentVariable}*");
-        }
-    }
-
-    public class TestVisualStudioExperimentalService : IExperimentationService
-    {
-        private readonly Dictionary<string, bool> _flights;
-
-        public TestVisualStudioExperimentalService()
-            : this(new Dictionary<string, bool>())
-        {
-        }
-
-        public TestVisualStudioExperimentalService(Dictionary<string, bool> flights)
-        {
-            _flights = flights ?? throw new ArgumentNullException(nameof(flights));
-        }
-
-        public void Dispose()
-        {
-            // do nothing
-        }
-
-        public bool IsCachedFlightEnabled(string flight)
-        {
-            _flights.TryGetValue(flight, out bool result);
-            return result;
-        }
-
-        public Task<bool> IsFlightEnabledAsync(string flight, CancellationToken token)
-        {
-            return Task.FromResult(IsCachedFlightEnabled(flight));
-        }
-
-        public void Start()
-        {
-            // do nothing.
-        }
-    }
-
-    public class TestOutputConsoleProvider : IOutputConsoleProvider
-    {
-        public TestOutputConsole TestOutputConsole { get; } = new TestOutputConsole();
-
-        public Task<IOutputConsole> CreateBuildOutputConsoleAsync()
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<IOutputConsole> CreatePackageManagerConsoleAsync()
-        {
-            return Task.FromResult((IOutputConsole)TestOutputConsole);
-        }
-
-        public Task<IConsole> CreatePowerShellConsoleAsync()
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class TestOutputConsole : IOutputConsole
-    {
-        public IList<string> Messages { get; } = new List<string>();
-
-        public int ConsoleWidth => throw new NotImplementedException();
-
-        public Task ActivateAsync()
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task ClearAsync()
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task WriteAsync(string text)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task WriteAsync(string text, Color? foreground, Color? background)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task WriteBackspaceAsync()
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task WriteLineAsync(string text)
-        {
-            Messages.Add(text);
-            return Task.CompletedTask;
-        }
-
-        public Task WriteLineAsync(string format, params object[] args)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task WriteProgressAsync(string currentOperation, int percentComplete)
-        {
-            throw new NotImplementedException();
         }
     }
 }
