+++ test/TestUtilities/Test.Utility/VisualStudio/NuGetExperimentationServiceUtility.cs
@@ -3,20 +3,125 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
+using System.Windows.Media;
+using Microsoft.VisualStudio.Experimentation;
 using NuGet.VisualStudio;
-using Moq;
 
 namespace Test.Utility.VisualStudio
 {
     public static class NuGetExperimentationServiceUtility
     {
-        public static Mock<INuGetExperimentationService> GetMock()
+        ////public static Mock<INuGetExperimentationService> GetMock()
+        //{
+        //    var mock = new Mock<INuGetExperimentationService>();
+        //    return mock;
+        //}
+    }
+
+    public class TestVisualStudioExperimentalService : IExperimentationService
+    {
+        private readonly Dictionary<string, bool> _flights;
+
+        public TestVisualStudioExperimentalService()
+            : this(new Dictionary<string, bool>())
         {
-            var mock = new Mock<INuGetExperimentationService>();
-            return mock;
+        }
+
+        public TestVisualStudioExperimentalService(Dictionary<string, bool> flights)
+        {
+            _flights = flights ?? throw new ArgumentNullException(nameof(flights));
+        }
+
+        public void Dispose()
+        {
+            // do nothing
+        }
+
+        public bool IsCachedFlightEnabled(string flight)
+        {
+            _flights.TryGetValue(flight, out bool result);
+            return result;
+        }
+
+        public Task<bool> IsFlightEnabledAsync(string flight, CancellationToken token)
+        {
+            return Task.FromResult(IsCachedFlightEnabled(flight));
+        }
+
+        public void Start()
+        {
+            // do nothing.
         }
     }
+
+    public class TestOutputConsoleProvider : IOutputConsoleProvider
+    {
+        public TestOutputConsole TestOutputConsole { get; } = new TestOutputConsole();
+
+        public Task<IOutputConsole> CreateBuildOutputConsoleAsync()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<IOutputConsole> CreatePackageManagerConsoleAsync()
+        {
+            return Task.FromResult((IOutputConsole)TestOutputConsole);
+        }
+
+        public Task<IConsole> CreatePowerShellConsoleAsync()
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    public class TestOutputConsole : IOutputConsole
+    {
+        public IList<string> Messages { get; } = new List<string>();
+
+        public int ConsoleWidth => throw new NotImplementedException();
+
+        public Task ActivateAsync()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task ClearAsync()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task WriteAsync(string text)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task WriteAsync(string text, Color? foreground, Color? background)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task WriteBackspaceAsync()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task WriteLineAsync(string text)
+        {
+            Messages.Add(text);
+            return Task.CompletedTask;
+        }
+
+        public Task WriteLineAsync(string format, params object[] args)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task WriteProgressAsync(string currentOperation, int percentComplete)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
 }
